:recycle: refactor: change unit test to unit test dir
