--- src/test/java/com/depromeet/streetdrop/common/annotation/MockAnonymousUser.java
@@ -1,15 +1,0 @@
-package com.depromeet.streetdrop.common.annotation;
-
-import org.springframework.security.test.context.support.TestExecutionEvent;
-import org.springframework.security.test.context.support.WithSecurityContext;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-@WithSecurityContext(factory = WithMockCustomAnonymousUserSecurityContextFactory.class, setupBefore = TestExecutionEvent.TEST_EXECUTION)
-public @interface MockAnonymousUser {
-
-    String name() default "koo";
-
-}
